# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zylora API",
    "description": "Managed multi-tenant websites, leads, messaging, SEO, and client operations.",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/billing/{tenant_id}/configuration": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Billing Configuration",
        "operationId": "get_billing_configuration_api_v1_billing__tenant_id__configuration_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Billing Configuration Api V1 Billing  Tenant Id  Configuration Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "billing"
        ],
        "summary": "Update Billing Configuration",
        "operationId": "update_billing_configuration_api_v1_billing__tenant_id__configuration_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingConfigurationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Billing Configuration Api V1 Billing  Tenant Id  Configuration Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/{tenant_id}/status": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Billing Status",
        "operationId": "get_billing_status_api_v1_billing__tenant_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Billing Status Api V1 Billing  Tenant Id  Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/{tenant_id}/clear-lockout": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Manual Clear Lockout",
        "operationId": "manual_clear_lockout_api_v1_billing__tenant_id__clear_lockout_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Manual Clear Lockout Api V1 Billing  Tenant Id  Clear Lockout Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/{tenant_id}/send-reminder": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Send Payment Reminder",
        "operationId": "send_payment_reminder_api_v1_billing__tenant_id__send_reminder_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Send Payment Reminder Api V1 Billing  Tenant Id  Send Reminder Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/{tenant_id}/pay-order": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Pay Now Order",
        "operationId": "create_pay_now_order_api_v1_billing__tenant_id__pay_order_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Pay Now Order Api V1 Billing  Tenant Id  Pay Order Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/{tenant_id}/verify-checkout": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Verify Pay Now Checkout",
        "operationId": "verify_pay_now_checkout_api_v1_billing__tenant_id__verify_checkout_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingCheckoutVerify"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Verify Pay Now Checkout Api V1 Billing  Tenant Id  Verify Checkout Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingCheckoutVerify": {
        "properties": {
          "razorpay_order_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 5,
            "title": "Razorpay Order Id"
          },
          "razorpay_payment_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 5,
            "title": "Razorpay Payment Id"
          },
          "razorpay_signature": {
            "type": "string",
            "maxLength": 512,
            "minLength": 8,
            "title": "Razorpay Signature"
          }
        },
        "type": "object",
        "required": [
          "razorpay_order_id",
          "razorpay_payment_id",
          "razorpay_signature"
        ],
        "title": "BillingCheckoutVerify"
      },
      "BillingConfigurationUpdate": {
        "properties": {
          "monthly_amount_minor": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Monthly Amount Minor"
          },
          "billing_currency": {
            "type": "string",
            "pattern": "^(USD|INR)$",
            "title": "Billing Currency"
          },
          "billing_day": {
            "type": "integer",
            "maximum": 31.0,
            "minimum": 1.0,
            "title": "Billing Day"
          }
        },
        "type": "object",
        "required": [
          "monthly_amount_minor",
          "billing_currency",
          "billing_day"
        ],
        "title": "BillingConfigurationUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}